+++ music_pipeline/tag_checker.py
@@ -153,4 +153,4 @@ def main(path=None):
         extract_and_display_images(audio)
 
 if __name__ == "__main__":
-    main("/home/adam/driveBig/Music/My Playlist")
+    main("D:/Music/My Playlist")